Expose helpers via export_test.go instead of an internal test package
Keep all tests in the json2hcl_test package and reach the unexported
helpers through export_test.go aliases.
