+++ unmarshal_error_test.go
@@ -1,6 +1,8 @@
 package json2hcl_test
 
 import (
+	"bytes"
+	"encoding/json"
 	"errors"
 	"testing"
 
@@ -45,4 +47,35 @@ func (errReader) Read([]byte) (int, error) {
 func TestUnmarshalFromError(t *testing.T) {
 	_, err := json2hcl.UnmarshalFrom(errReader{})
 	assert.Error(t, err)
+}
+
+// The following cover defensive branches that the public API can't reach,
+// because encoding/json rejects the corresponding inputs before they get this
+// far. They exercise the existing code as-is rather than reshaping it.
+
+func TestScalarTokensError(t *testing.T) {
+	assert := assert.New(t)
+
+	// json.Number always holds a valid number when produced by the decoder,
+	// so this parse failure is only reachable by calling scalarTokens directly.
+	_, err := json2hcl.ScalarTokens(json.Number("not-a-number"))
+	assert.Error(err)
+
+	// The decoder never yields a token of this type.
+	_, err = json2hcl.ScalarTokens(123)
+	assert.Error(err)
+}
+
+func TestDecodeValueUnexpectedDelim(t *testing.T) {
+	assert := assert.New(t)
+
+	dec := json.NewDecoder(bytes.NewReader([]byte("{}")))
+
+	_, err := dec.Token() // consume the opening '{'
+	assert.NoError(err)
+
+	// The next token is the closing '}', a delimiter decodeValue never expects
+	// in a value position.
+	_, err = json2hcl.DecodeValue(dec)
+	assert.Error(err)
 }
